+++ codex/signals.py
@@ -5,7 +5,15 @@ import signal
 from asyncio import Event
 from logging import getLogger
 
+from django.core.cache import cache
 from django.db.backends.signals import connection_created
+from django.db.models.signals import m2m_changed
+
+from codex.librarian.queue_mp import (
+    LIBRARIAN_QUEUE,
+    BroadcastNotifierTask,
+    DelayedTasks,
+)
 
 
 LOG = getLogger(__name__)
@@ -40,13 +48,27 @@ def bind_signals(loop):
     loop.add_signal_handler(signal.SIGUSR1, _restart_signal_handler)
 
 
-def _activate_wal_journal(sender, connection, **kwargs):
+def _activate_wal_journal(sender, connection, **kwargs):  # noqa: F841
     """Enable sqlite WAL journal."""
     with connection.cursor() as cursor:
         cursor.execute("PRAGMA journal_mode=wal;")
         LOG.debug("sqlite journal_mode=wal")
 
 
+def _user_group_change(action, instance, pk_set, model, **kwargs):  # noqa: F841
+    """Clear cache and send update signals when groups change."""
+    if model.__name__ == "Group" and action in (
+        "post_add",
+        "post_remove",
+        "post_clear",
+    ):
+        cache.clear()
+        tasks = (BroadcastNotifierTask("LIBRARY_CHANGED"),)
+        task = DelayedTasks(2, tasks)
+        LIBRARIAN_QUEUE.put(task)
+
+
 def connect_signals():
     """Connect actions to signals."""
     connection_created.connect(_activate_wal_journal)
+    m2m_changed.connect(_user_group_change)