+++ docker/docker-version-codex-arch.sh
@@ -1,10 +1,9 @@
 #!/bin/bash
 # Get the final runnable codex image version
 set -euo pipefail
-source .env
-VERSION=${PKG_VERSION}
+VERSION=$(./version.sh)
 if [ "${CIRCLECI:-}" ]; then
-    ARCH=$(./docker-arch.sh)
+    ARCH=$(./docker/docker-arch.sh)
     VERSION=${VERSION}-${ARCH}
 fi
 echo "$VERSION"