+++ lint-backend.sh
@@ -30,7 +30,7 @@ if [ "$(uname)" = "Darwin" ]; then
     shfmt -d -i 4 ./*.sh ./**/*.sh
     # subdirs aren't copied into docker builder
     # .env files aren't copied into docker
-    shellcheck --external-sources ./**/*.sh .env .env.platforms .env.pushover
+    shellcheck --external-sources ./**/*.sh .env.platforms .env.pushover
     circleci config check .circleci/config.yml
 fi
 shellcheck --external-sources ./*.sh